Stable initialization

+++ gpytorch/variational/cholesky_variational_distribution.py
@@ -35,7 +35,9 @@ class CholeskyVariationalDistribution(VariationalDistribution):
 
     def initialize_variational_distribution(self, prior_dist):
         self.variational_mean.data.copy_(prior_dist.mean)
-        self.chol_variational_covar.data.copy_(prior_dist.covariance_matrix.inverse().cholesky())
+        self.chol_variational_covar.data.copy_(
+            prior_dist.covariance_matrix.double().inverse().cholesky().type_as(prior_dist.mean)
+        )
 
     @property
     def variational_distribution(self):
